# Slightly improve parsing speed