This is a small patch that does the following:

 - Adds a "hint" to tolit that avoids needing to loop to retrieve
   literals
 - Replaces (slow) modulos with bitwise functions
 - Modified number parsing to be more compact and faster
 - Most notably, replaces copying string() function in ParseBytes with a
   dangerous but faster path. Since gjson uses strings internally and in
   return values and cannot modify this, this is safe [citation needed]

This improves performance between 0 and 10%:

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkGJSONGet                 621           577           -7.09%
BenchmarkGJSONGetMany4Paths       524           489           -6.68%
BenchmarkGJSONGetMany8Paths       361           340           -5.82%
BenchmarkGJSONGetMany16Paths      274           259           -5.47%
BenchmarkGJSONGetMany32Paths      219           212           -3.20%
BenchmarkGJSONGetMany64Paths      201           195           -2.99%
BenchmarkGJSONGetMany128Paths     187           182           -2.67%
BenchmarkGJSONUnmarshalMap        6972          6935          -0.53%
BenchmarkGJSONUnmarshalStruct     6606          6128          -7.24%
```