@tidwall 2017-09-02
Hi Connor,

Thanks for the PR. Some of this looks promising. I'll review it in depth later in the week and let you know if I find any problems.

Thanks!