@tidwall 2017-09-09
Hey Connor, 

Sorry about the delay. 

Excellent work. I just reviewed and everything looks good with the exception of the unsafe code.

The problem is that the new unsafe code opens up the possibility for the Result strings to be change by modifying the original json `[]byte` slice.

For example:
```go
json := []byte(`{"name":"Tom"}`)
r := gjson.ParseBytes(json)
println(r.Raw)
json[len(json)-3] = 'd'
println(r.Raw)
```
Prints
```
Tom
Tod
```

So we should probably revert that portion of the PR. The rest I'll merge.

Thanks!