@tidwall 2017-11-13
I'm closing for now. Feel free to reopen when you fix the unsafe bytes issue. Thanks.